requirements.txt
fastapi[standard]==0.124.4
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
pymongo==4.16.0